Dodan textview u user profilu koji prikazuje korisnikov rank. Rank se dohvaca kad se na main menu-u dohvaca profil usera i sprema u Json.file

+++ webservice/src/main/java/com/example/webservice/interfaces/WebServiceCaller.java
@@ -114,7 +114,7 @@ public class WebServiceCaller {
         CheckCall();
     }
 
-    public void getMainMenu(float longitude, float latitude, int radius){
+    public void getMainMenu(double longitude, double latitude, int radius){
         callWorkspaces = serviceCaller.getMainMenu(longitude,latitude,radius);
         CheckWorkspaceCall();
     }
